Renamed package for FailoverConsumerApp to `com.example` from `app2`
