+++ settings.gradle
@@ -18,7 +18,7 @@ include 'spring-pulsar-cache-provider-caffeine'
 include 'spring-pulsar-reactive'
 include 'spring-pulsar-dependencies'
 include 'spring-pulsar-sample-apps:sample-imperative-produce-consume'
-include 'spring-pulsar-sample-apps:sample-app2'
+include 'spring-pulsar-sample-apps:sample-failover-custom-router'
 include 'spring-pulsar-sample-apps:sample-pulsar-functions:sample-signup-app'
 include 'spring-pulsar-sample-apps:sample-pulsar-functions:sample-signup-function'
 include 'spring-pulsar-sample-apps:sample-reactive'
