+++ spring-pulsar-sample-apps/sample-failover-custom-router/build.gradle
