+++ spring-pulsar-sample-apps/sample-failover-custom-router/compose.yaml
