+++ spring-pulsar-sample-apps/sample-failover-custom-router/src/main/java/com/example/package-info.java
@@ -3,7 +3,7 @@
  */
 @NonNullApi
 @NonNullFields
-package app2;
+package com.example;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
